Change IPC prefix from bitcoin: to dogecoin:
Given that GUIUtil::parseBitcoinURI is expecting it to start with dogecoin,
it seems like this was overlooked, and this results in dogecoin: links not
being processed as expected in handleURIOrFile, ipcParseCommandLine.

Instead of processing the link as a payment request and opening the such send
page, it simply opens the core wallet (or attempts to open another).

This commit makes dogecoin-qt properly handle dogecoin links so that the such
send page is opened with the desired information obtained from the link.

Fixes #1628