Fix warning in dbwrapper test
While `buf` isn't going to overflow with the values written to it, the compiler
can't prove that, because the integers used for the contents of the buffer
could possibly be too large.

Fortunately, these integers only need a couple of bits.

`uint8_t` would be enough but it might be better to be more consistent with
other types in this function.